api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
databases
pydantic[email]
python-dotenv
requests
